Move LanguageWrapper to i18n; remove unused references to it; wrap return values from html render in LanguageWrapper
